Refactor firmware to State Pattern and modularize web interface
Firmware Refactoring:
* Implemented State Pattern for display modes to decouple logic from the main loop.
* Introduced DisplayMode base class and concrete implementations:
    * TimeMode: Handles clock functionality.
    * DateMode: Handles date display.
    * ManualMode: Direct character control.
    * MultiInputMode: Multi-character input handling.
    * SingleInputMode: Single-character input handling.
    * RandomTestMode: Testing functionality.
* Moved configuration constants to src/Defaults.h.
* Extracted helper functions to src/Utils.h.
* Updated SplitFlapDisplay.ino to use the new DisplayMode system.
* Updated SplitFlapModule classes to align with the new structure.

Web Interface Modularization:
* Split monolithic index.js into modular files in src/web/js/:
    * main.js: Entry point.
    * settings.js: Configuration management.
    * controls.js: Display control logic.
    * calibration.js: Calibration routines.
    * common.js: Shared utilities.
* Updated settings.html to work with the new module structure.
* Updated vite.config.mjs to handle the new file organization.

Documentation:
* Added RECOMMENDATIONS.md outlining the improvements.
