+++ platformio.ini
@@ -35,7 +35,7 @@ build_flags=
 [env:ota]
 upload_protocol=espota
 upload_port=splitflap.local
-; upload_flags = --auth=yourpassword
+upload_flags = --auth=flap234@
 
 [env:esp32_c3]
 extends=env
