+++ src/SplitFlapDisplay.h
@@ -15,30 +15,97 @@
 
 class SplitFlapMqtt;
 
+/**
+ * @brief Main controller class for the Split Flap Display.
+ * 
+ * Manages multiple SplitFlapModule instances, handles high-level display operations
+ * (writing strings, homing), and coordinates with MQTT.
+ */
 class SplitFlapDisplay {
   public:
     SplitFlapDisplay(JsonSettings &settings);
 
+    /**
+     * @brief Initialize the display and all modules.
+     */
     void init();
-    void updateOffsets();  // Update offsets without full reinit
+
+    /**
+     * @brief Update module offsets from settings without full re-initialization.
+     */
+    void updateOffsets();
+
+    /**
+     * @brief Write a string to the display.
+     * 
+     * @param inputString The string to display
+     * @param speed Motor speed in RPM
+     * @param centering Whether to center the text if shorter than display width
+     */
     void writeString(
         String inputString, float speed = MAX_RPM,
         bool centering = true
-    );                                     // Move all modules at once to show a specific string
+    );
+
+    /**
+     * @brief Set all modules to a single character.
+     * 
+     * @param inputChar The character to display
+     * @param speed Motor speed in RPM
+     */
     void writeChar(char inputChar,
-                   float speed = MAX_RPM); // sets all modules to a single char
+                   float speed = MAX_RPM);
+
+    /**
+     * @brief Move modules to specific step positions.
+     * 
+     * @param targetPositions Array of target positions for each module
+     * @param speed Motor speed in RPM
+     * @param releaseMotors Whether to de-energize motors after movement
+     * @param isHoming Whether this movement is part of a homing sequence
+     */
     void moveTo(int targetPositions[], float speed = MAX_RPM, bool releaseMotors = true, bool isHoming = false);
-    void home(float speed = MAX_RPM);      // move home
+
+    /**
+     * @brief Home all modules (find zero position).
+     * 
+     * @param speed Motor speed in RPM
+     */
+    void home(float speed = MAX_RPM);
+
+    /**
+     * @brief Home all modules and then display a string.
+     * 
+     * @param homeString The string to display after homing
+     * @param speed Motor speed in RPM
+     * @param centering Whether to center the text
+     */
     void homeToString(
         String homeString, float speed = MAX_RPM,
         bool centering = true
-    );                                      // moves home and then writes a string
+    );
+
+    /**
+     * @brief Home all modules and then display a character.
+     * 
+     * @param homeChar The character to display after homing
+     * @param speed Motor speed in RPM
+     */
     void homeToChar(char homeChar,
-                    float speed = MAX_RPM); // moves home and then sets all modules to a char
+                    float speed = MAX_RPM);
+
     void testAll();
     void testCount();
     void testRandom(float speed = MAX_RPM);
-    void testModule(int moduleIndex, float speed = MAX_RPM); // Test single module: A -> 0 -> blank
+
+    /**
+     * @brief Test a single module (cycle A -> 0 -> blank).
+     * 
+     * @param moduleIndex Index of the module to test
+     * @param speed Motor speed in RPM
+     */
+    void testModule(int moduleIndex, float speed = MAX_RPM);
+
     int getNumModules() { return numModules; }
     int getCharsetSize() const { return charSetSize; }
     void setMqtt(SplitFlapMqtt *mqttHandler);
