+++ src/Utils.h
@@ -1,0 +1,60 @@
+#pragma once
+#include <Arduino.h>
+#include <time.h>
+
+inline String trimToModuleCount(const String &str, int maxLen) {
+    return str.length() > maxLen ? str.substring(0, maxLen) : str;
+}
+
+inline String extractFromCSV(String str, int index) {
+    int startIndex = 0;
+    int endIndex = str.length();
+
+    int commaCount = 0;
+    for (int i = 0; i < str.length(); i++) {
+        if (str[i] == ',') {
+            commaCount++;
+            if (commaCount == index) {
+                startIndex = i + 1; // skip past the comma
+            } else if (commaCount == index + 1) {
+                endIndex = i;
+            }
+        }
+    }
+
+    return str.substring(startIndex, endIndex);
+}
+
+inline String convertToStrftime(String userFormat) {
+    struct FormatToken
+    {
+        const char *token;
+        const char *strftime;
+    };
+
+    FormatToken tokens[] = {
+        // Date formats
+        {"{yyyy}", "%Y"}, // 4-digit year (e.g. 2025)
+        {"{dddd}", "%A"}, // Full weekday name (e.g. Monday)
+        {"{mmmm}", "%B"}, // Full month name (e.g. January)
+        {"{ddd}", "%a"},  // Abbreviated weekday name (e.g. Mon)
+        {"{mmm}", "%b"},  // Abbreviated month name (e.g. Apr)
+        {"{dd}", "%d"},   // 2-digit day of month, zero-padded (01–31)
+        {"{mm}", "%M"},   // Minutes (00–59)
+        {"{yy}", "%y"},   // 2-digit year (e.g. 25)
+        {"{ww}", "%V"},   // ISO 8601 week number (01–53)
+        {"{D}", "%j"},    // Day of the year (001–366)
+
+        // Time formats
+        {"{HH}", "%H"},   // Hours (24-hour clock, 00–23)
+        {"{hh}", "%I"},   // Hours (12-hour clock, 01–12)
+        {"{MM}", "%m"},   // 2-digit month number, zero-padded (01–12)
+        {"{AMPM}", "%p"}, // AM or PM
+    };
+
+    for (auto &t : tokens) {
+        userFormat.replace(t.token, t.strftime);
+    }
+
+    return userFormat;
+}
