+++ src/web/index.html
@@ -11,7 +11,7 @@
             }
         </style>
         <link rel="stylesheet" href="index.css" />
-        <script defer src="index.js" type="module"></script>
+        <script type="module" src="./js/main.js"></script>
     </head>
     <body
         x-data="page('Control')"
